Next.js 13.4 practice
this repo is just for practice purposes, trying to learn the new Next.js framework version 13.4, using the new /App Route directory as well as /src directory. Also trying out new technologies, such as Prisma, Typescript and Tailwind CSS (custom HTML components, custom Tailwind classes and colors), as well as implementing Auth0 and many more dependencies englobing a large project. Note: THIS REPOSITORY IS A PRODUCT OF ANOTHER PERSON'S WORK AND SHOULD NOT BE REGARDED AS MY OWN. ALSO THIS REPOSITORY IS NOT MEANT TO BE A FULLY FLEDGED PROJECT, AS IT IS PRACTICE AND SHOULD BE REGARDED AS AN INCOMPLETE WORK